+++ source/main/cardano/setup.js
@@ -49,6 +49,7 @@ const startCardanoNode = (
     syncTolerance,
     cliBin,
     isStaging,
+    smashUrl,
   } = launcherConfig;
   const logFilePath = `${logsPrefix}/pub/`;
   const config = {
@@ -65,6 +66,7 @@ const startCardanoNode = (
     syncTolerance,
     cliBin,
     isStaging,
+    smashUrl,
     startupTimeout: NODE_STARTUP_TIMEOUT,
     startupMaxRetries: NODE_STARTUP_MAX_RETRIES,
     shutdownTimeout: NODE_SHUTDOWN_TIMEOUT,